Test using a different hasher won't compile
Currently, the only public way to construct a `BloomFilterBuilder`
is by using `BloomFilterBuilder::default()`, which only constructs
a `BloomFilterBuilder<RandomState, CompressedBitmap>`.

Then the `hasher` function takes a `hasher` of type `H`, which can
only ever be called with a `RandomState` as that's the only kind of
`BloomFilterBuilder` you can have.

So you can have a BloomFilterBuilder of any hasher you want, as long as
it's RandomState!

Thus, this test fails to compile:

```
error[E0308]: mismatched types
   --> src/bloom.rs:446:21
    |
446 |             .hasher(BuildHasherDefault::<twox_hash::XxHash64>::default())
    |              ------ ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ expected `RandomState`, found `BuildHasherDefault<XxHash64>`
    |              |
    |              arguments to this method are incorrect
    |
    = note: expected struct `std::hash::RandomState`
               found struct `BuildHasherDefault<XxHash64>`
```