feat: write-optimised VecBitmap
This commit adds the VecBitmap storage implementation of the Bitmap
trait, allowing it to be used as the storage backing of the bloom
filter.

The CompressedBitmap is read and space optimised, with that space gain
having to trade off write performance as a result. This VecBitmap makes
the reverse trade - O(n) space for O(1) writes.