feat: convert VecBitmap to CompressedBitmap
This commit enables an uncompressed, write optimised VecBitmap to be
converted into a read-and-space-optimised CompressedBitmap.

This improves performance of a common use case: write a bunch of values
into a filter at the start, and then perform mostly (or only) reads
after.

Now a filter can be bulk loaded with the VecBitmap and then reduced to a
CompressedBitmap for later reads.