build: bump to 0.5.0
Adds an uncompressed VecBitmap backing store for the bitmap to enable
optimal write performance for bulk loading, and then later convertible
into the CompressedBitmap to optimise for reads & memory usage.

This release contains a breaking change - the
BloomFilterBuilder::with_bitmap() method was renamed to
with_bitmap_data() (and unsafe annotation removed) and a new method was
added with using the old name.